New Elastic Rope Joint Def
Everythin is based on Distance Joints just modified omega and frequency
parameters depending on 2 cases.

If the new length is higher than the set length frequency and damping
is used to create a stretch effect.
If the new Length is smaller than the set length alls position and
velocity corrections are skipped.